fix(#1261): break NestJS module cycle from catalog→playlist discovery
CatalogModule importing PlaylistModule closed a cycle
(SharedModule → GenerationModule → CatalogModule → PlaylistModule →
SharedModule), so Nest could not construct PlaylistModule and the whole app
failed to boot — red Backend Integration, Backend Tests, and E2E (the E2E web
server timed out). Local tests missed it because they used isolated test
modules / `new PlaylistService()` and never booted the full AppModule.

Fix: keep listPublicPlaylists in PlaylistService and serve GET /catalog/playlists
from a new PublicPlaylistDiscoveryController mounted in PlaylistModule under the
`catalog/playlists` path. Same public URL, no CatalogModule→PlaylistModule
import, no cycle. Reverted the catalog controller/module changes and their test
deltas; added a focused unit + HTTP spec for the new controller. Verified the
full app now boots (health.integration green) alongside the discovery and
public-playlist integration suites.
