+++ backend/src/modules/catalog/catalog.module.ts
@@ -3,10 +3,9 @@ import { CatalogController } from "./catalog.controller";
 import { CatalogService } from "./catalog.service";
 import { EncryptionModule } from "../encryption/encryption.module";
 import { RightsModule } from "../rights/rights.module";
-import { PlaylistModule } from "../playlist/playlist.module";
 
 @Module({
-  imports: [EncryptionModule, RightsModule, PlaylistModule],
+  imports: [EncryptionModule, RightsModule],
   controllers: [CatalogController],
   providers: [CatalogService],
   exports: [CatalogService],
