+++ backend/src/modules/playlist/playlist.module.ts
@@ -1,11 +1,11 @@
 import { Module } from "@nestjs/common";
-import { PlaylistController, PublicPlaylistController } from "./playlist.controller";
+import { PlaylistController, PublicPlaylistController, PublicPlaylistDiscoveryController } from "./playlist.controller";
 import { PlaylistService } from "./playlist.service";
 import { SharedModule } from "../shared/shared.module";
 
 @Module({
     imports: [SharedModule],
-    controllers: [PlaylistController, PublicPlaylistController],
+    controllers: [PlaylistController, PublicPlaylistController, PublicPlaylistDiscoveryController],
     providers: [PlaylistService],
     exports: [PlaylistService],
 })
